feat: add seo
