+++ src/app/(home)/(info)/about/page.tsx
@@ -1,10 +1,42 @@
+import { APP_DESCRIPTION, APP_NAME, APP_URL } from '@/app/constants'
 import { Button } from '@/components/ui/button'
 import { ArrowRight, Bot, Box, ShieldCheck, Globe } from 'lucide-react'
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+	metadataBase: new URL(APP_URL),
+
+	title: 'About Us',
+	description: APP_DESCRIPTION,
+
+	openGraph: {
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		url: APP_URL,
+		siteName: APP_NAME,
+		images: [
+			{
+				url: '/og-image.png',
+				width: 1200,
+				height: 630,
+				alt: `Preview image for ${APP_NAME}`,
+			},
+		],
+		locale: 'en_US',
+		type: 'website',
+	},
+
+	twitter: {
+		card: 'summary_large_image',
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		images: ['/og-image.png'],
+	},
+}
 
 export default function AboutPage() {
 	const features = [
 		{
-			// Иконки теперь используют цвета, адаптированные для темы
 			icon: <Bot className='h-8 w-8 text-zinc-900 dark:text-white' />,
 			title: 'AI-Powered Generation',
 			description:
