+++ src/app/(home)/(info)/enterprise/page.tsx
@@ -1,3 +1,4 @@
+import { APP_DESCRIPTION, APP_NAME, APP_URL } from '@/app/constants'
 import { BackgroundRippleEffect } from '@/components/ui/background-ripple-effect'
 import { Button } from '@/components/ui/button'
 import {
@@ -12,7 +13,39 @@ import {
 	Code,
 	MessageCircle,
 } from 'lucide-react'
+import { Metadata } from 'next'
 import Link from 'next/link'
+
+export const metadata: Metadata = {
+	metadataBase: new URL(APP_URL),
+
+	title: 'Enterprise',
+	description: APP_DESCRIPTION,
+
+	openGraph: {
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		url: APP_URL,
+		siteName: APP_NAME,
+		images: [
+			{
+				url: '/og-image.png',
+				width: 1200,
+				height: 630,
+				alt: `Preview image for ${APP_NAME}`,
+			},
+		],
+		locale: 'en_US',
+		type: 'website',
+	},
+
+	twitter: {
+		card: 'summary_large_image',
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		images: ['/og-image.png'],
+	},
+}
 
 const enterpriseFeatures = [
 	{
