+++ src/app/(home)/layout.tsx
@@ -1,5 +1,54 @@
+import type { Metadata } from 'next'
 import { Footer } from '@/modules/home/ui/components/footer'
 import { Navbar } from '@/modules/home/ui/components/navbar/navbar'
+import { APP_DESCRIPTION, APP_NAME, APP_URL } from '@/app/constants'
+
+export const metadata: Metadata = {
+	metadataBase: new URL(APP_URL),
+
+	title: {
+		default: APP_NAME,
+		template: `${APP_NAME} | %s`,
+	},
+	description: APP_DESCRIPTION,
+
+	icons: {
+		icon: [
+			{
+				url: '/logo.svg',
+				media: '(prefers-color-scheme: light)',
+			},
+			{
+				url: '/logo-l.svg',
+				media: '(prefers-color-scheme: dark)',
+			},
+		],
+	},
+
+	openGraph: {
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		url: APP_URL,
+		siteName: APP_NAME,
+		images: [
+			{
+				url: '/og-image.png',
+				width: 1200,
+				height: 630,
+				alt: `Preview image for ${APP_NAME}`,
+			},
+		],
+		locale: 'en_US',
+		type: 'website',
+	},
+
+	twitter: {
+		card: 'summary_large_image',
+		title: APP_NAME,
+		description: APP_DESCRIPTION,
+		images: ['/og-image.png'],
+	},
+}
 
 interface Props {
 	children: React.ReactNode
