+++ src/app/constants/index.ts
@@ -1,3 +1,4 @@
 export const APP_NAME = 'Lummie'
 export const APP_DESCRIPTION =
 	'Build your dream application with our powerful AI tools and components.'
+export const APP_URL = 'https://lummie.dev'
