+++ src/app/projects/[projectId]/page.tsx
@@ -1,19 +1,44 @@
 import { ErrorBoundary } from 'react-error-boundary'
 import { ProjectView } from '@/modules/projects/ui/views/project-view'
-import { getQueryClient, trpc } from '@/trpc/server'
+import { getQueryClient, trpc, caller } from '@/trpc/server'
 import { dehydrate, HydrationBoundary } from '@tanstack/react-query'
 import { Suspense } from 'react'
+import { Metadata } from 'next'
+
+export async function generateMetadata({ params }: Props): Promise<Metadata> {
+	const { projectId } = params
+
+	try {
+		const project = await caller.projects.getOne({ id: projectId })
+
+		if (!project) {
+			return {
+				title: 'Project Not Found',
+			}
+		}
+
+		return {
+			title: `${project.name ?? 'Generating...'} | Lummie`,
+		}
+	} catch {
+		return {
+			title: 'Error',
+			description: 'Could not fetch project data.',
+		}
+	}
+}
 
 interface Props {
-	params: Promise<{
+	params: {
 		projectId: string
-	}>
+	}
 }
 
 const Page = async ({ params }: Props) => {
 	const { projectId } = await params
 
 	const queryClient = getQueryClient()
+
 	void queryClient.prefetchQuery(
 		trpc.messages.getMany.queryOptions({ projectId })
 	)
@@ -23,7 +48,6 @@ const Page = async ({ params }: Props) => {
 
 	return (
 		<HydrationBoundary state={dehydrate(queryClient)}>
-			{/* need to create error page */}
 			<ErrorBoundary fallback={<div>Something went wrong</div>}>
 				<Suspense fallback={<div>Loading...</div>}>
 					<ProjectView projectId={projectId} />
